+++ src/index.js
@@ -1,7 +1,9 @@
 import Info from "./Info";
+import Success from "./Success";
 import Warning from "./Warning";
 
 export {
   Info,
+  Success,
   Warning
 };